Reducing index.js
